+++ lib/charter_agreement_protocol/limits.ex
@@ -9,14 +9,22 @@ defmodule CharterAgreementProtocol.Limits do
 
   alias CharterAgreementProtocol.Error
 
-  @fields [:max_bytes, :max_depth, :max_object_members, :max_array_items, :max_string_bytes]
+  @fields [
+    :max_bytes,
+    :max_depth,
+    :max_object_members,
+    :max_array_items,
+    :max_string_bytes,
+    :max_artifact_set_items
+  ]
 
   @defaults [
     max_bytes: 1_048_576,
     max_depth: 64,
     max_object_members: 1_024,
     max_array_items: 4_096,
-    max_string_bytes: 65_536
+    max_string_bytes: 65_536,
+    max_artifact_set_items: 1_024
   ]
 
   @maximums %{
@@ -24,7 +32,8 @@ defmodule CharterAgreementProtocol.Limits do
     max_depth: 128,
     max_object_members: 65_536,
     max_array_items: 65_536,
-    max_string_bytes: 1_048_576
+    max_string_bytes: 1_048_576,
+    max_artifact_set_items: 4_096
   }
 
   @enforce_keys @fields
@@ -35,7 +44,8 @@ defmodule CharterAgreementProtocol.Limits do
           max_depth: non_neg_integer(),
           max_object_members: non_neg_integer(),
           max_array_items: non_neg_integer(),
-          max_string_bytes: non_neg_integer()
+          max_string_bytes: non_neg_integer(),
+          max_artifact_set_items: non_neg_integer()
         }
 
   @doc "The closed limit-field set."